+++ src/OpenSearch.Client/Modules/SnapshotAndRestore/Repositories/GetRepository/GetRepositoryResponseFormatter.cs
@@ -29,6 +29,7 @@
 using System;
 using System.Collections.Generic;
 using OpenSearch.Net;
+using OpenSearch.Net.Extensions;
 using OpenSearch.Net.Utf8Json;
 using OpenSearch.Net.Utf8Json.Resolvers;